add InitConfigurator class
