+++ src/input/AxisConfigWidget.cc
@@ -32,6 +32,8 @@
 #include "input/InputDriverManager.h"
 #include "SettingsWriter.h"
 #include "WheelIgnorer.h"
+#include "initConfig.h"
+
 
 AxisConfigWidget::AxisConfigWidget(QWidget *parent) : QWidget(parent)
 {
@@ -48,7 +50,7 @@ void AxisConfigWidget::AxesChanged(int nr, double val) const{
 
 	int value = val * 100;
 	progressBar->setValue(value); //set where the bar is
-
+	
 	//QProgressBar generates the shown string from the format string.
 	//By setting a format string without a place holder,
 	//we can set arbitrary text, like a custom formatted double.
@@ -511,75 +513,12 @@ void AxisConfigWidget::applyComboBox(QComboBox *comboBox, int val, Settings::Set
 	writeSettings();
 }
 
-void AxisConfigWidget::initSpinBox(QSpinBox *spinBox, const Settings::SettingsEntry& entry)
-{
-	RangeType range = entry.range().toRange();
-	spinBox->setMinimum(range.begin_value());
-	spinBox->setMaximum(range.end_value());
-}
-
-void AxisConfigWidget::initDoubleSpinBox(QDoubleSpinBox *spinBox, const Settings::SettingsEntry& entry)
-{
-	Settings::Settings *s = Settings::Settings::inst();
-
-	RangeType range = entry.range().toRange();
-	spinBox->blockSignals(true);
-	spinBox->setMinimum(range.begin_value());
-	spinBox->setSingleStep(range.step_value());
-	spinBox->setMaximum(range.end_value());
-	spinBox->setValue((double)s->get(entry).toDouble());
-	spinBox->blockSignals(false);
-}
-
-void AxisConfigWidget::initCheckBox(QCheckBox *checkBox, const Settings::SettingsEntry& entry)
-{
-	Settings::Settings *s = Settings::Settings::inst();
-	const Value &value = s->get(entry);
-	bool state = value.toBool();
-
-	checkBox->blockSignals(true);
-	checkBox->setChecked(state);
-	checkBox->blockSignals(false);
-}
-
-void AxisConfigWidget::updateComboBox(QComboBox *comboBox, const Settings::SettingsEntry& entry)
-{
-	Settings::Settings *s = Settings::Settings::inst();
-
-	const Value &value = s->get(entry);
-	QString text = QString::fromStdString(value.toString());
-	int idx = comboBox->findData(text);
-	if (idx >= 0) {
-		comboBox->setCurrentIndex(idx);
-	} else {
-		const Value &defaultValue = entry.defaultValue();
-		QString defaultText = QString::fromStdString(defaultValue.toString());
-		int defIdx = comboBox->findData(defaultText);
-		if (defIdx >= 0) {
-			comboBox->setCurrentIndex(defIdx);
-		} else {
-			comboBox->setCurrentIndex(0);
-		}
-	}
-}
-
 void AxisConfigWidget::writeSettings()
 {
 	SettingsWriter settingsWriter;
 	Settings::Settings::inst()->visit(settingsWriter);
 }
 
-void AxisConfigWidget::initComboBox(QComboBox *comboBox, const Settings::SettingsEntry& entry)
-{
-	comboBox->clear();
-	// Range is a vector of 2D vectors: [[name, value], ...]
-	for(const auto &v : entry.range().toVector()) {
-		QString val = QString::fromStdString(v[0]->toString());
-		QString qtext = QString::fromStdString(gettext(v[1]->toString().c_str()));
-		comboBox->addItem(qtext, val);
-	}
-	updateComboBox(comboBox, entry);
-}
 
 void AxisConfigWidget::updateStates(){
 	if(!initialized) return;
